Add context support for h5parm class (#145)
Add context support to `h5parm` class

An `h5parm` object can now also be used in a with context. To retain backward compatibility, the underlying HDF5 file object is still opened in the `__init__()` method, but is now closed in the `__exit__()` method, which is implicitly called when leaving the `with` context.

Existing code will continue to function as it used to. You still need to call `close()` explicitly to close the HDF5 file object. New code should preferably use the h5parm object only inside a `with` context, thus ensuring that the HDF5 file object will always be closed, even when an exception is raised and not caught.