Replace rcond with rtol in pinv() call (#156)
* Replace rcond with rtol in pinv() call

The `pinv()` function deprecated the use of `rcond` in favour of `rtol`, and SciPy 1.14.0 removed the keyword option `rcond` completely from `pinv()`.
Replaced all occurrences of `rcond` in calls to `pinv()` with `rtol`.

* Update python.yml

Need to fix workflow as well, since it still used the old-fashioned way of installing (i.e. `python setup.py install`, instead of `pip install .`)

* Do not use NumPy 2.0

Package `python-casacore` is not (yet) compatible with NumPy 2.0. Use latest NumPy 1.x version.

* Update python.yml

Wrong tests were called.

Just invoking `pytest` should do the right thing. To be fixed in another PR.

* We cannot yet use PyTest

The name of the (single) test script doesn't adhere to `pytest` standards, so it is not found.
At the same time, there are a few Python files named "test*.py" that contain currently failing tests.
This needs to be fixed in a different PR.

* Update python.yml

Use latest major version of `actions/setup-python`