Use pyproject.toml instead of setup.py (#168)
* Use raw string in regular expression

* Replace setup.py with pyproject.toml

* Bump Python version in GitHub workflow

* Drop Python 3.13 for now

* LoSoTo is not yet compatible with Numpy 2.x

* Automatically generate version number

Use `setuptools_scm` to automatically determine the version number.
The H5Parm version number is now defined in `__init__.py`

* Do not put all files in source distribution

Setuptools, when combined with setuptools_scm, puts all the files that are under version control in the source distribution. We want to be a bit moreselective. With this MANIFEST.in file we exactly mimic the old behaviour.