Make LoSoto compatible with NumPy 2.x (#169)
* Add testing with tox

Added testing with tox to the `pyproject.toml` file.

* Specify dependencies for testing in pyproject.toml

Removed the `Install dependencies` job from the GitHub workflow. The dependencies are needed for testing are now specified as `[project.optional-dependencies]` in `pyproject.toml`.

* Clean up workflow

Removed now obsolete lines of code in GitHub action.

* Make code Numpy 2.x compatible

So far, it looks like only one change needed to be made. The obsolete alias `numpy.string_` has been replaced with `numpy.bytes`, as suggested in the error message (and in the NumPy 2.0 release notes: https://numpy.org/doc/stable/release/2.0.0-notes.html#numpy-2-0-python-api-removals).

* Use Numpy 1.x with Python 3.9

Not completely clear why, but `numpy` 2.x causes issues with the `tables` package when using Python 3.9.
So, let's restrict `numpy` to 1.x when using Python 3.9.

* Add Python 3.13 to the test

Now that LoSoTo is Numpy 2.x compatible, it should be possible to test with Python 3.13 as well.