Update tests
Updated Losoto operation tests such that each skeleton test calls the `run` function with exactly the same arguments as used in the definition of the `run` function in the operation itself.
Note that some tests fail with a `NameError` exception, because mandatory arguments for some of the `run` functions have not (yet) been defined in the test(s). Also note that the current fixture used to create a `soltab` doesn't contain the expected data for some of the operations, which therefore correctly fail.