Clean up code tree
Removed directory `losoto/tests`. It didn't contain any (working) tests we'd want to keep, except for `losoto_test.py`, which was moved to the `tests` directory, because we want to retain it (at least temporarily).
Removed some out-dated files in the `tools` directory. Maybe it can go completely, but I'm not sure (yet).