Setup test framework (#172)
* Setup test files/directories

Created a directory structure and populated it with test files. All files are still empty, so there's nothing to test yet.

* Implement minimal test for H5Parm

Implemented a minimal suite of tests (consisting of _one_ test) for the `losoto.h5parm` module

* Add extra skeleton test files

Added few, still empty, skeleton files for testing the modules in the losoto directory.

* Add placeholder test functions

Added empty placeholders for testing the main classes and functions in the losoto module.

* Add tests that actually test operations

Added skeleton tests, but at the moment they all fail, because the input to the function that creates a Solset is failing on missing inputs (which are not mandatory, BTW).

* Add some basic tests for each Losoto operation

This is still work-in-progress. Need to think more about which fixtures are needed.

* Remove spurious pass statements

Since all test functions now at least contain some code, the `pass` statements have been removed.

* Add action to report test coverage

Experiment a bit with some test coverage reporting tools.

* Continue pipeline even if some tests fail

* Also run pipeline on pull request

* Update tests

Updated Losoto operation tests such that each skeleton test calls the `run` function with exactly the same arguments as used in the definition of the `run` function in the operation itself.
Note that some tests fail with a `NameError` exception, because mandatory arguments for some of the `run` functions have not (yet) been defined in the test(s). Also note that the current fixture used to create a `soltab` doesn't contain the expected data for some of the operations, which therefore correctly fail.

* Clean up code tree

Removed directory `losoto/tests`. It didn't contain any (working) tests we'd want to keep, except for `losoto_test.py`, which was moved to the `tests` directory, because we want to retain it (at least temporarily).
Removed some out-dated files in the `tools` directory. Maybe it can go completely, but I'm not sure (yet).

* Give h5 fixture more descriptive name