Add sample comparison output
The file `compare.out` contains the difference between two runs.
Note that the run times may vary with a few seconds between runs.
The differences are not something to be worried about, though for very short processes (e.g. in the PLOT operation) you'll notice that the new implementation has a bit more overhead.